[IMP] certificate, l10n_es_edi_*: Factor patched HTTPAdapter for certificates models
We generalize the CertificateAdapter that is needed to make zeep read
our own Certificate model for the session.

task-4477745

X-original-commit: 303718627af35405b550ee41d16918666f6577eb
