+++ addons/certificate/tools/__init__.py
@@ -1,0 +1,1 @@
+from .certificate_adapter import CertificateAdapter
